+++ src/constructor/__init__.py
@@ -1,0 +1,4 @@
+from .registry import Registry
+
+OPTIMIZERS = Registry('optimizers')
+SCHEDULERS = Registry('schedulers')
