--- src/registry/__init__.py
@@ -1,4 +1,0 @@
-from .registry import Registry
-
-OPTIMIZERS = Registry('optimizers')
-SCHEDULERS = Registry('schedulers')
